+++ api/services/user/store_getuserbyid.go
@@ -2,6 +2,7 @@ package user
 
 import (
 	"context"
+	"fmt"
 
 	"github.com/MicroFish91/portfolio-instruments-api/api/constants"
 	"github.com/MicroFish91/portfolio-instruments-api/api/types"
@@ -13,14 +14,14 @@ func (s *PostgresUserStore) GetUserById(ctx context.Context, userId int) (types.
 
 	row := s.db.QueryRow(
 		c,
-		`
+		fmt.Sprintf(`
 			select
-				*
+				%s
 			from
 				users
 			where
 				user_id = $1
-		`,
+		`, userColumns),
 		userId,
 	)
