+++ api/services/user/store_getusers.go
@@ -2,10 +2,10 @@ package user
 
 import (
 	"context"
+	"fmt"
 
 	"github.com/MicroFish91/portfolio-instruments-api/api/constants"
 	"github.com/MicroFish91/portfolio-instruments-api/api/types"
-	"github.com/jackc/pgx/v5"
 )
 
 func (s *PostgresUserStore) GetUsers(ctx context.Context, options *types.GetUsersStoreOptions) ([]types.User, types.PaginationMetadata, error) {
@@ -29,9 +29,9 @@ func (s *PostgresUserStore) GetUsers(ctx context.Context, options *types.GetUser
 
 	rows, err := s.db.Query(
 		c,
-		`
+		fmt.Sprintf(`
 			select
-				*, 
+				%s, 
 				COUNT(*) OVER() as total_items
 			from
 				users
@@ -43,7 +43,7 @@ func (s *PostgresUserStore) GetUsers(ctx context.Context, options *types.GetUser
 				$1
 			offset
 				$2	
-		`,
+		`, userColumns),
 		options.Page_size,
 		(options.Current_page-1)*options.Page_size,
 	)
@@ -63,30 +63,4 @@ func (s *PostgresUserStore) GetUsers(ctx context.Context, options *types.GetUser
 		Page_size:    options.Page_size,
 		Total_items:  totalItems,
 	}, nil
-}
-
-func (s *PostgresUserStore) parseRowsIntoUsers(rows pgx.Rows) ([]types.User, int, error) {
-	var users []types.User
-	var total_items int
-
-	for rows.Next() {
-		var u types.User
-		err := rows.Scan(
-			&u.User_id,
-			&u.Email,
-			&u.Enc_password,
-			&u.User_role,
-			&u.Last_logged_in,
-			&u.Verified,
-			&u.Created_at,
-			&u.Updated_at,
-			&total_items,
-		)
-
-		if err != nil {
-			return []types.User{}, 0, nil
-		}
-		users = append(users, u)
-	}
-	return users, total_items, nil
 }
